+++ rs/tests/src/rosetta_tests/rosetta_client.rs
@@ -471,6 +471,27 @@ impl RosettaApiClient {
         )
     }
 
+    pub async fn get_known_neurons(&self) -> Result<Result<CallResponse, Error>, String> {
+        let req = CallRequest::new(
+            ic_rosetta_api::models::NetworkIdentifier(self.network_id()),
+            "list_known_neurons".to_owned(),
+            ObjectMap::new(),
+        );
+        debug!(&self.logger, "[Rosetta client] Call Request: {:?}", req);
+        debug!(
+            &self.logger,
+            "[Rosetta client] Call: {}",
+            &format!("{}/call", self.api_url)
+        );
+        to_rosetta_response::<CallResponse>(
+            self.post_json_request(
+                &format!("{}/call", self.api_url),
+                serde_json::to_vec(&req).unwrap(),
+            )
+            .await,
+        )
+    }
+
     pub async fn get_pending_proposals(&self) -> Result<Vec<Proposal>, String> {
         let req = CallRequest::new(
             ic_rosetta_api::models::NetworkIdentifier(self.network_id()),